+++ launcher/cli/apps.go
@@ -1083,9 +1083,10 @@ func init() {
 		RegisterFlags: func(cmd *cobra.Command) error {
 			cmd.Flags().String("booter-bootseq", "", "File path tp the desired boot sequence")
 			cmd.Flags().String("booter-nodeos-api-addr", NodeosAPIAddr, "Target API address to communicate with underlying nodeos")
-			cmd.Flags().String("booter-cache-path", "{dfuse-data-dir}/booter/cache", "Directory to store downloaded contract and ABI data")
+			cmd.Flags().String("booter-data-dir", "{dfuse-data-dir}/booter", "Booter's working directory")
 			cmd.Flags().String("booter-vault-file", "", "Wallet file that contains encrypted key material")
 			cmd.Flags().String("booter-private-key", "", "Genesis private key")
+
 			return nil
 		},
 		FactoryFunc: func(modules *launcher.RuntimeModules) (launcher.App, error) {
@@ -1097,7 +1098,7 @@ func init() {
 			return boot.New(&boot.Config{
 				NodeosAPIAddress: viper.GetString("booter-nodeos-api-addr"),
 				BootSeqFile:      viper.GetString("booter-bootseq"),
-				CachePath:        mustReplaceDataDir(dfuseDataDir, viper.GetString("booter-cache-path")),
+				Datadir:          mustReplaceDataDir(dfuseDataDir, viper.GetString("booter-data-dir")),
 				VaultPath:        viper.GetString("booter-vault-file"),
 				PrivateKey:       viper.GetString("booter-private-key"),
 			}), nil